+++ packages/sp-clientsvc/src/objectpath.ts
@@ -239,29 +239,6 @@ export class ObjectPathQueue {
         return this._contextIndex;
     }
 
-    /**
-     * Generates a ~unique hash code for this ObjectPathQueue
-     * 
-     * From: https://stackoverflow.com/questions/6122571/simple-non-secure-hash-function-for-javascript
-     */
-    // tslint:disable:no-bitwise
-    public hash(): number {
-        const s = this.toBody();
-        let hash = 0;
-        if (s.length === 0) {
-            return hash;
-        }
-
-        for (let i = 0; i < s.length; i++) {
-            const chr = s.charCodeAt(i);
-            hash = ((hash << 5) - hash) + chr;
-            hash |= 0; // Convert to 32bit integer
-        }
-        return hash;
-
-    }
-    // tslint:enable:no-bitwise
-
     public toBody(): string {
 
         if (objectDefinedNotNull(this._xml)) {